+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Inventory causes an error. Refs UIIN-2012.
 * Fix Name Type facet sends requests to incorrect endpoint. Refs UIIN-2062.
 * Browse Contributors: Hitting Next/Previous has odd behavior. Refs UIIN-2058
 * Fix filters list doesn't update when user back to "Browse contributors" pane. Refs UIIN-2069.
+* Create Jest/RTL test for PrimaryToggleButton. Refs UIIN-1666.
 
 ## [9.0.0](https://github.com/folio-org/ui-inventory/tree/v9.0.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v8.0.0...v9.0.0)